flask/requirements.txt
flask
gtts
requests
pillow
numpy
ultralytics
paho-mqtt
pygame 